Revolver Minor Updates
