+++ addons/TAS_Revolver/CfgAmmo.hpp
@@ -85,8 +85,8 @@ class CfgAmmo
 	};
 	class TAS_ammo_doomsday_buckshot: TAS_ammo_m576_buckshot
 	{
-		hit = 10; //originally 9 (for previous doomsday before buff)
-		indirectHit = 10; //originally 9
+		hit = 12; //originally 9 (for previous doomsday before buff)
+		indirectHit = 8; //originally 9
 		indirectHitRange = 0.6;
 	};
 
@@ -108,8 +108,8 @@ class CfgAmmo
 	// HE FRAG
 	class TAS_ammo_12g_FRAG : B_12Gauge_Slug
 	{
-		hit = 7; //originally 6 before buff
-		indirectHit = 1.5; //original 1.143
+		hit = 9; //originally 6 before buff
+		indirectHit = 2; //original 1.143
 		indirectHitRange = 2;
 		caliber = 0.2;
 		cartridge = "";
